requirements.txt
requests==2.32.3
python-dotenv==1.0.1
jinja2==3.1.4
playwright==1.48.0
pytest==8.3.3
flask==3.0.3
